Restrict users to only update or delete their own tasks

+++ controller/tasksController.js
@@ -1,7 +1,7 @@
 import pool from '../config.js';
 
 const getTasks = (req, res) => {
-  console.log('Header', req.headers['authorization'])
+  console.log(req.user.id)
   pool
     .query('SELECT * FROM tasks ORDER BY id DESC')
     .then(data => res.status(200).json(data.rows))
@@ -25,18 +25,22 @@ const searchCatOrStat = (req, res) => {
 };
 
 const updateTask = (req, res) => {
-  const {id, user_id, category, description, status} = req.body; 
+  const { id, user_id, category, description, status } = req.body;
+  console.log(req.user.id)
+  if (!(user_id === req.user.id)) return res.status(400).json({ Error: 'Every user can update their own tasks' });
   pool
-  .query(`UPDATE tasks set category = $1, description = $2, status=$3 WHERE id=$4 RETURNING *`, [category, description, status, id])
-  .then(data => res.status(200).json({message: 'To do item updated'}))
-  .catch(err => res.status(502).json({ error: err.message}))
+    .query(`UPDATE tasks set category = $1, description = $2, status=$3 WHERE id=$4 RETURNING *`, [category, description, status, id])
+    .then(data => res.status(200).json({ message: 'To do item updated' }))
+    .catch(err => res.status(502).json({ error: err.message }))
 };
+
 const deleteTask = (req, res) => {
-  const {id, user_id} = req.body; 
+  const { id, user_id } = req.body;
+  if (!(user_id === req.user.id)) return res.status(400).json({ Error: 'Every user can delete their own tasks' });
   pool
-  .query(`DELETE FROM tasks WHERE id=$1 RETURNING *`, [id])
-  .then(data => res.status(200).json({message: 'To do item deleted'}))
-  .catch(err => res.status(502).json({ error: err.message}))
+    .query(`DELETE FROM tasks WHERE id=$1 RETURNING *`, [id])
+    .then(data => res.status(200).json({ message: 'To do item deleted' }))
+    .catch(err => res.status(502).json({ error: err.message }))
 };
 
 export default { getTasks, createTask, updateTask, deleteTask, searchCatOrStat };
